backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic>=2.7.0,<3.0.0
python-multipart==0.0.6
youtube-transcript-api==0.6.1
pytube==15.0.0
openai>=1.55.3
python-dotenv==1.0.0
requests==2.31.0 
firebase-admin==6.4.0
stripe==8.5.0